+++ index.js
@@ -5,6 +5,7 @@ module.exports = {
     "no-getState": require("./rules/no-getState/no-getState"),
     "no-unnecessary-duplication": require("./rules/no-unnecessary-duplication/no-unnecessary-duplication"),
     "prefer-sample-over-forward-with-mapping": require("./rules/prefer-sample-over-forward-with-mapping/prefer-sample-over-forward-with-mapping"),
+    "no-duplicate-clock-or-source-array-values": require("./rules/no-duplicate-clock-or-source-array-values/no-duplicate-clock-or-source-array-values"),
     "no-useless-methods": require("./rules/no-useless-methods/no-useless-methods"),
     "no-ambiguity-target": require("./rules/no-ambiguity-target/no-ambiguity-target"),
     "no-watch": require("./rules/no-watch/no-watch"),
@@ -17,11 +18,13 @@ module.exports = {
     "no-guard": require("./rules/no-guard/no-guard"),
     "mandatory-scope-binding": require("./rules/mandatory-scope-binding/mandatory-scope-binding"),
     "prefer-useUnit": require("./rules/prefer-useUnit/prefer-useUnit"),
+    "no-patronum-debug": require("./rules/no-patronum-debug/no-patronum-debug"),
   },
   configs: {
     recommended: require("./config/recommended"),
     scope: require("./config/scope"),
     react: require("./config/react"),
     future: require("./config/future"),
+    patronum: require("./config/patronum"),
   },
 };