+++ rules/enforce-store-naming-convention/enforce-store-naming-convention.js
@@ -1,3 +1,5 @@
+const { ESLintUtils } = require("@typescript-eslint/utils");
+
 const { extractImportedFrom } = require("../../utils/extract-imported-from");
 const { namingOf } = require("../../utils/naming");
 const {
@@ -32,12 +34,17 @@ module.exports = {
     hasSuggestions: true,
   },
   create(context) {
-    const { parserServices } = context;
+    let parserServices;
+    try {
+      parserServices = ESLintUtils.getParserServices(context);
+    } catch (err) {
+      // no types information
+    }
 
     validateStoreNameConvention(context);
 
     // TypeScript-way
-    if (parserServices.hasFullTypeInformation) {
+    if (parserServices?.program) {
       return {
         VariableDeclarator(node) {
           const isEffectorStore = nodeTypeIs.store({
@@ -76,19 +83,22 @@ module.exports = {
         // Store creation with method
         const STORE_CREATION_METHODS = ["createStore", "restore", "combine"];
         for (const method of STORE_CREATION_METHODS) {
-          const localMethod = importedFromEffector.get(method);          
+          const localMethod = importedFromEffector.get(method);
           if (!localMethod) {
             continue;
           }
-            
+
           const isEffectorStoreCreation = node.callee.name === localMethod;
           if (!isEffectorStoreCreation) {
             continue;
           }
 
-          const parentNode = traverseParentByType(node, "VariableDeclarator", ["Program"]);
+          const parentNode = traverseParentByType(node, "VariableDeclarator", {
+            stopOnTypes: ["Program", "BlockStatement"],
+          });
 
-          const resultSavedInVariable = parentNode.type === "VariableDeclarator";
+          const resultSavedInVariable =
+            parentNode?.type === "VariableDeclarator";
           if (!resultSavedInVariable) {
             continue;
           }
@@ -116,7 +126,8 @@ module.exports = {
             return;
           }
 
-          const resultSavedInVariable = node.parent.type === "VariableDeclarator";
+          const resultSavedInVariable =
+            node.parent.type === "VariableDeclarator";
 
           if (!resultSavedInVariable) {
             return;
@@ -141,10 +152,12 @@ module.exports = {
         if (
           STORE_IN_DOMAIN_CREATION_METHODS.includes(node.callee?.property?.name)
         ) {
+          const parentNode = traverseParentByType(node, "VariableDeclarator", {
+            stopOnTypes: ["Program", "BlockStatement"],
+          });
 
-          const parentNode = traverseParentByType(node, "VariableDeclarator", ["Program"]);
-
-          const resultSavedInVariable = parentNode.type === "VariableDeclarator";
+          const resultSavedInVariable =
+            parentNode?.type === "VariableDeclarator";
           if (!resultSavedInVariable) {
             return;
           }