+++ utils/create-link-to-rule.js
@@ -1,5 +1,5 @@
 function createLinkToRule(name) {
-  return `https://github.com/effector/eslint-plugin/blob/master/rules/${name}/${name}.md`;
+  return `https://eslint.effector.dev/rules/${name}`;
 }
 
 module.exports = { createLinkToRule };