+++ utils/node-type-is.js
@@ -1,7 +1,10 @@
+const { ESLintUtils } = require("@typescript-eslint/utils");
+
 function hasType({ node, possibleTypes, context, from }) {
   try {
-    const checker = context.parserServices.program.getTypeChecker();
-    const originalNode = context.parserServices.esTreeNodeToTSNodeMap.get(node);
+    const parserServices = ESLintUtils.getParserServices(context);
+    const checker = parserServices.program.getTypeChecker();
+    const originalNode = parserServices.esTreeNodeToTSNodeMap.get(node);
     const type = checker.getTypeAtLocation(
       originalNode?.initializer ?? originalNode
     );
@@ -21,13 +24,27 @@ const nodeTypeIs = {
   effect: (opts) =>
     hasType({ ...opts, possibleTypes: ["Effect"], from: "effector" }),
   store: (opts) =>
-    hasType({ ...opts, possibleTypes: ["Store"], from: "effector" }),
+    hasType({
+      ...opts,
+      possibleTypes: ["Store", "StoreWritable"],
+      from: "effector",
+    }),
   event: (opts) =>
-    hasType({ ...opts, possibleTypes: ["Event"], from: "effector" }),
+    hasType({
+      ...opts,
+      possibleTypes: ["Event", "EventCallable"],
+      from: "effector",
+    }),
   unit: (opts) =>
     hasType({
       ...opts,
-      possibleTypes: ["Effect", "Store", "Event"],
+      possibleTypes: [
+        "Effect",
+        "Store",
+        "Event",
+        "EventCallable",
+        "StoreWritable",
+      ],
       from: "effector",
     }),
   gate: (opts) =>
@@ -55,13 +72,27 @@ const nodeTypeIs = {
         from: "effector",
       }),
     store: (opts) =>
-      !hasType({ ...opts, possibleTypes: ["Store"], from: "effector" }),
+      !hasType({
+        ...opts,
+        possibleTypes: ["Store", "StoreWritable"],
+        from: "effector",
+      }),
     event: (opts) =>
-      !hasType({ ...opts, possibleTypes: ["Event"], from: "effector" }),
+      !hasType({
+        ...opts,
+        possibleTypes: ["Event", "EventCallable"],
+        from: "effector",
+      }),
     unit: (opts) =>
       !hasType({
         ...opts,
-        possibleTypes: ["Effect", "Store", "Event"],
+        possibleTypes: [
+          "Effect",
+          "Store",
+          "Event",
+          "EventCallable",
+          "StoreWritable",
+        ],
         from: "effector",
       }),
     gate: (opts) =>